feat(client): promote Stores folder to higher parent directory
