+++ client/src/stores/BatchStore.ts
@@ -1,8 +1,8 @@
 import { asyncDerived } from '@square/svelte-store';
-import { assert } from '../../../assert.ts';
+import { assert } from '../assert.ts';
 import { topToastMessage } from './ToastStore.ts';
 import { dashboardState } from './DashboardState.ts';
-import { Batch } from '../../../api/batch.ts';
+import { Batch } from '../api/batch.ts';
 
 export const earliestBatch = asyncDerived(
     dashboardState,
