+++ client/src/stores/InviteStore.ts
@@ -1,15 +1,15 @@
 import { asyncDerived } from '@square/svelte-store';
+import { assert } from '../assert.ts';
 import { dashboardState } from './DashboardState.ts';
-import { Staff } from '../../../api/staff.ts';
-import { assert } from '../../../assert.ts';
 import { topToastMessage } from './ToastStore.ts';
+import { Invite } from '../api/invite.ts';
 
-export const staffList = asyncDerived(
+export const inviteList = asyncDerived(
     dashboardState,
     ({ currentOffice }) => {
         try {
             if (currentOffice !== null)
-                return Staff.get(currentOffice);
+                return Invite.getList(currentOffice);
         } catch (err) {
             assert(err instanceof Error);
             topToastMessage.enqueue({ title: err.name, body: err.message });
