+++ client/src/stores/OfficeStore.ts
@@ -1,6 +1,6 @@
 import { asyncReadable } from '@square/svelte-store';
 
-import { Office } from '../../../api/office.ts';
+import { Office } from '../api/office.ts';
 
 export const allOffices = asyncReadable(
     { },
